+++ ksa_compliance/ksa_compliance/doctype/zatca_business_settings/zatca_business_settings.py
@@ -309,7 +309,7 @@ class ZATCABusinessSettings(Document):
 
     @staticmethod
     def for_invoice(
-        invoice_id: str, doctype: Literal["Sales Invoice", "POS Invoice"]
+        invoice_id: str, doctype: Literal["Sales Invoice", "POS Invoice", "Payment Entry"]
     ) -> Optional["ZATCABusinessSettings"]:
         company_id = frappe.db.get_value(doctype, invoice_id, ["company"])
         if not company_id: